fix(golang): move server to golang
